services/evals/pyproject.toml
[project]
name = "ragbench-evals"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "pyyaml>=6.0",
    "llama-index-core>=0.14.4",
    "llama-index-llms-ollama>=0.8.3",
    "llama-index-llms-openai>=0.3.33",
    "llama-index-llms-anthropic>=0.6.10",
    "llama-index-llms-google-genai>=0.1.0",
    "llama-index-llms-deepseek>=0.1.0",
    "anthropic>=0.75.0",
    "datasets>=3.2.0",
    "deepeval>=3.6.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]
